+++ front/src/styles/Aoferta.css
@@ -2,15 +2,24 @@
     margin:0;
     padding: 0;
 }
+.contenedor{
+    background-color: rgb(27, 27, 27);
+}
 
-.card {
+.cardsita {
+    border-radius: 10px;
+    margin-left: 10px !important;
+}
+.cabecera{
+    background: #53b7f1;
     border-radius: 10px;
 }
-.btn{
+.pagar{
     position: relative;
     margin-left: 80px;
+    background-color: #ffc200;
 }
-section{
+.section{
     position: relative;
     width: 100%;
     height: 100vh;
@@ -19,20 +28,24 @@ section{
     align-items: center;
     padding: 100px;
     box-sizing: border-box;
-    background: linear-gradient(180deg, #1899cc, #ffc200), url(../Assets/img2.jpg);
+    background: linear-gradient(180deg, #4156cc, #ffc200), url(../assets/img1.jpg);
+    background-size: cover;
+    background-position: right;
+
+    background-repeat: no-repeat;
     background-blend-mode: screen;
 }
-section .content{
+.section .content{
     max-width: 700px;
     text-align: center;
 }
-section .content h2{
+.section .content h2{
     font-size: 3em;
-    color: #fff;
+    color: rgb(10, 10, 10);
 }
-section .content p{
+.section .content p{
     font-size: 1.2em;
-    color: #fff;
+    color: rgb(51, 49, 49);
 }
 .wave{
     position: absolute;
